# resources/swagger/georchestra-signalement.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Georchestra - Plugin Signalement API",
		"contact": {
			"email": "guillaume.morin@open-groupe.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "",
	"basePath": "/signalement",
...
	"schemes": [
		"https",
		"http"
	],
	"paths": {
...
		"/users": {
			"get": {
				"tags": [
					"user"
				],
				"summary": "Recherche de la liste des utilisateurs.",
				"description": "Récupération de la liste des utilisateurs disponibles qui respectent les filtres.",
				"operationId": "searchUsers",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"description": "Filtre de recherche sur l'e-mail",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "loginUser",
						"description": "Filtre de recherche sur le login",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer",
						"default": 0
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer",
						"default": 10
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string",
						"default": ""
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/UserPageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"post": {
				"tags": [
					"user"
				],
				"summary": "Pré création d'un utilisateur.",
				"description": "Pré création d'un utilisateur.",
				"operationId": "createUser",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "Le User à précréé",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/User"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/users/{login}": {
			"get": {
				"tags": [
					"user"
				],
				"summary": "Récupération d'un utilisateur",
				"description": "Récupération d'un utilisateur en fonction de son login",
				"operationId": "getUser",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "login",
						"description": "Login de l'utilisateur'",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/User"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"user"
				],
				"summary": "Suppression d'un utilisateur",
				"description": "Suppression d'un utilisateur en fonction de son login",
				"operationId": "deleteUser",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "login",
						"description": "Login de l'utilisateur'",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/userRoleContexts": {
			"get": {
				"tags": [
					"userRoleContext"
				],
				"summary": "Recherche des userRoleContext",
				"description": "Récupération de la liste des userRoleContext disponibles qui respectent les filtres.",
				"operationId": "searchUserRoleContexts",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "userLogin",
						"description": "Filtre de recherche sur le user",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "roleName",
						"description": "Filtre de recherche sur le role",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "geographicAreaId",
						"description": "Filtre de recherche sur la zone géographique",
						"required": false,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "query",
						"name": "contextDescriptionName",
						"description": "Filtre de recherche sur le contexte",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer",
						"default": 0
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer",
						"default": 10
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string",
						"default": ""
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/UserRoleContextPageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"post": {
				"tags": [
					"userRoleContext"
				],
				"summary": "Création d'un opérateur.",
				"description": "Création d'un opérateur.",
				"operationId": "createUserRoleContext",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "Les infos sur le UserRoleContext à créer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserRoleContext"
						}
					}
				],
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/UserRoleContext"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/userRoleContexts/{id}": {
			"get": {
				"tags": [
					"userRoleContext"
				],
				"summary": "Récupération d'un opérateur",
				"description": "Récupération d'un opérateur en fonction de son id",
				"operationId": "getUserRoleContext",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "id de l'opérateur",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/UserRoleContext"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"userRoleContext"
				],
				"summary": "Suppression d'un opérateur",
				"description": "Suppression d'un opérateur en fonction de son id",
				"operationId": "deleteUserRoleContext",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "id de l'opérateur",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "query",
						"name": "force",
						"description": "Force la suppression si l'opérateur a des signalements en cours",
						"required": false,
						"type": "boolean",
						"default": false
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		}
	},
	"definitions": {
		"ApiError": {
			"type": "object",
			"properties": {
				"code": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"GeographicType": {
			"type": "string",
			"description": "Type géographic",
			"enum": [
				"POINT",
				"LINE",
				"POLYGON"
			]
		},
...
		"ContextType": {
			"type": "string",
			"description": "Type du context du signalement",
			"enum": [
				"THEMA",
				"LAYER"
			]
		},
...
		"ContextDescription": {
			"type": "object",
			"properties": {
				"contextType": {
					"type": "object",
					"$ref": "#/definitions/ContextType"
				},
				"geographicType": {
					"type": "object",
					"$ref": "#/definitions/GeographicType"
				},
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"processDefinitionKey": {
					"type": "string"
				},
				"revision": {
					"type": "integer"
				}
			}
		},
...
		"User": {
			"type": "object",
			"properties": {
				"login": {
					"type": "string"
				},
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"organization": {
					"type": "string"
				},
				"roles": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
		"UserPageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/User"
					}
				}
			}
		},
...
		"GeographicArea": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"nom": {
					"type": "string"
				},
				"codeInsee": {
					"type": "string"
				}
			}
		},
...
		"Role": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"UserRoleContext": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"user": {
					"type": "object",
					"$ref": "#/definitions/User"
				},
				"role": {
					"type": "object",
					"$ref": "#/definitions/Role"
				},
				"contextDescription": {
					"type": "object",
					"$ref": "#/definitions/ContextDescription"
				},
				"geographicArea": {
					"type": "object",
					"$ref": "#/definitions/GeographicArea"
				}
			}
		},
		"UserRoleContextPageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/UserRoleContext"
					}
				}
			}
		}
	}
}